Desktop/recc/requirements.txt
fastapi
uvicorn[standard]
neo4j
graphdatascience
python-dotenv
